# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tresleai Rest API ",
    "description": "Tresleai Facade Microservice API",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1.0/history/retrieval": {
      "get": {
        "tags": [
          "crate"
        ],
        "summary": "GET handler to extract a specific document from an application's history collection, with an input 'reference_id' as the basis for retrieval.",
        "description": "This API is designed to be used after initiating a retrieval process via the retrieval API. The retrieval API returns a 'reference_id' which is then used as the basis for retrieving the corresponding document from the history collection of an application within the database. The document contains the initiated query and its generated response, among other relevant details.\n\n#### API Key\n- The application's API key is required to authenticate the request.\n- This API key is created during the application onboarding process and is persisted in the API gateway of the concerned AWS account.\n- It must be included in the `x-api-key` header of the request to associate it with an application.\n\n#### Example\n\nConsider an application, named as \"test_app\", equipped with an API key \"a8VYYvaey38pajBi4jrMt8pGNdw5w0pn8oCytuQB\". This application has previously executed a retrieval operation, and obtained a reference ID \"14b1456d-2708-45bc-8989-eac2d2eba4db\" from the retrieval API.\nThe application can now extract the document associated with this specific reference ID from its history collection.\n\nTo accomplish this, a GET request would be made to the below endpoint, with the `reference_id` as a query parameter:\n\n```\nGET /api/v1.0/history/retrieval?reference_id=\"14b1456d-2708-45bc-8989-eac2d2eba4db\"\nx-api-key: a8VYYvaey38pajBi4jrMt8pGNdw5w0pn8oCytuQB\n```\n\nUpon successful retrieval of the document, the response would be returned as follows:\n\n```\n{\n\"status\": \"success\",\n\"message\": \"History document with reference ID: \"14b1456d-2708-45bc-8989-eac2d2eba4db\" retrieved successfully.\",\n\"app_name\": \"test_app\"\n\"data\": {\n\"_id\": {\n\"$oid\": \"\u003Cid of the document in DocumentDB\u003E\"\n},\n\"reference_id\": \"14b1456d-2708-45bc-8989-eac2d2eba4db\",\n\"task_id\": \"\u003Ctask_id\u003E\",\n\"query\": \"\u003Cquery\u003E\",\n\"response\": \"\u003Cresponse\u003E\",\n\"timestamp\": \"\u003Ctimestamp\u003E\"\n}\n}\n```\n\nPlease note that the document is created in the database only upon successful generation of the response.\nUntil this point, the document is in the 'processing' state, indicated by a 202 (ACCEPTED) status code,\nas demonstrated in the following example response (including a sample reference ID):\n\n```\n{\n\"message\": \"Retrieval in progress.\",\n\"reference_id\": \"0103f9a3-bbd7-40a6-9545-4122d63fa0a8\",\n\"error_code\": 202\n}\n```",
        "operationId": "get_history_handler",
        "parameters": [
          {
            "name": "reference_id",
            "in": "query",
            "description": "Reference id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History document retrieved successfully."
          },
          "202": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          },
          "400": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          },
          "404": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          },
          "500": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          }
        }
      }
    },
    "/api/v1.0/retrieval": {
      "post": {
        "tags": [
          "crate"
        ],
        "summary": "POST handler to initiate the retrieval process for a response corresponding to a specific query.",
        "description": "This API triggers a retrieval operation in the backend, which fetches data based on the user details and query provided in the request. The retrieval process involves validating user policies, and asynchronously initiating the response retrieval by passing on the query and user details to the engine.\n\nThe request comprises several important components:\n\n#### User Details\n- The 'user_details' field contains the user's information, inclusive of their ID and IAM policy particulars.\n- The 'access_details' field holds the details of the IAM policy and database policy.\n- The IAM policies tied to the user outline their permissions and access rights to resources in the AWS environment. These policies are validated before proceeding with retrieval, else the request is terminated.\n- The database policy details include the name of the database and table.\n- Each policy is linked with a unique name and ARN (Amazon Resource Name).\n- For instance, a policy might allow the user access to certain S3 buckets, or grant permissions to operate on other AWS resources. This would shape a tailored response based on the resources the user can access.\n\n#### Query and additional prompt\n- The 'query' field contains the query to initiate the retrieval.\n- For enhanced context, the 'additional_prompt' field can be utilized.\n\n#### API Key\n- The application's API key is required to authenticate the request.\n- This API key is created during the application onboarding process and is persisted in the API gateway of the concerned AWS account.\n- It must be included in the `x-api-key` header of the request to associate it with an application.\n\n#### Example\n\nConsider a user with ID \"sample_user@example.com\", with access granted through two policies: \"policy1\" (\"arn:aws:iam::aws:policy/policy1\"), and \"policy2\" (\"arn:aws:iam::aws:policy/policy2\"). The application's API key is \"a8VYYvaey38pajBi4jrMt8pGNdw5w0pn8oCytuQB\". The user wants to initiate a retrieval with the query \"provide a list of all accessible documents\" and an additional prompt \"related to policy1\".\n\nTo initiate retrieval, a POST request would be made to the below endpoint:\n\n```\nPOST /api/v1.0/retrieval\nx-api-key: a8VYYvaey38pajBi4jrMt8pGNdw5w0pn8oCytuQB\n```\n\nBelow is an example of how the request payload might look like:\n\n```\n{\n\"user_details\": {\n\"user_id\": \"sample_user@example.com\",\n\"access_details\": {\n\"iam_policy_details\": [\n{\n\"policy_name\": \"policy1\",\n\"policy_arn\": \"arn:aws:iam::aws:policy/policy1\"\n},\n{\n\"policy_name\": \"policy2\",\n\"policy_arn\": \"arn:aws:iam::aws:policy/policy2\"\n}\n]\n\"db_policy_details\": [\n{\n\"database_name\": \"database1\",\n\"table_name\": \"table1\"\n},\n{\n\"database_name\": \"database2\",\n\"table_name\": \"table2\"\n}\n]\n}\n},\n\"query\": \"provide a list of all accessible documents\",\n\"additional_prompt\": \"related to policy1\"\n}\n```\n\nUpon successful initiation, the response would be returned as follows:\n\n```\n{\n\"status\": \"success\",\n\"message\": \"Retrieval initiated successfully.\",\n\"reference_id\": \"14b1456d-2708-45bc-8989-eac2d2eba4db\"\n}\n```\n\nIn the above response, the returned reference ID is then used to call the history retrieval API to fetch the response document from the database. It can also be used to contact Tresle support team with any questions or concerns.",
        "operationId": "post_retrieval_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Retrieval in progress."
          },
          "400": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          },
          "404": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          },
          "500": {
            "description": "Internal Error. Please contact tresleai support team. Use reference ID: "
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessDetails": {
        "type": "object",
        "properties": {
          "db_policy_details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DbPolicyDetails"
            },
            "nullable": true
          },
          "iam_policy_details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IAMPolicyDetails"
            },
            "nullable": true
          }
        }
      },
...
      "DbPolicyDetails": {
        "type": "object",
        "required": [
          "database_name",
          "table_name"
        ],
        "properties": {
          "database_name": {
            "type": "string"
          },
          "table_name": {
            "type": "string"
          }
        }
      },
...
      "IAMPolicyDetails": {
        "type": "object",
        "required": [
          "policy_name",
          "policy_arn"
        ],
        "properties": {
          "policy_arn": {
            "type": "string"
          },
          "policy_name": {
            "type": "string"
          }
        }
      },
...
      "RetrievalRequest": {
        "type": "object",
        "required": [
          "user_details",
          "query",
          "additional_prompt"
        ],
        "properties": {
          "additional_prompt": {
            "type": "string"
          },
          "app_name": {
            "type": "string",
            "nullable": true
          },
          "query": {
            "type": "string"
          },
          "task_id": {
            "type": "string",
            "nullable": true
          },
          "user_details": {
            "$ref": "#/components/schemas/UserDetails"
          }
        }
      },
...
      "UserDetails": {
        "type": "object",
        "required": [
          "user_id",
          "access_details"
        ],
        "properties": {
          "access_details": {
            "$ref": "#/components/schemas/AccessDetails"
          },
          "user_id": {
            "type": "string"
          }
        }
      }
    }
  }
}